+++ src/__tests__/Api/updateScore.spec.js
@@ -1,4 +1,4 @@
-import updateScores from '../js/Api/updateScore.js';
+import updateScores from '../../js/Api/updateScore.js';
 
 require('jest-fetch-mock').enableMocks();
