+++ src/js/Scenes/GameScene.js
@@ -108,6 +108,7 @@ export default class GameScene extends Phaser.Scene {
     this.scoreText.setText(`Score: ${this.sys.game.globals.score}`);
   }
 
+
   async onMeetEnemy(player, zone) {
     zone.x = Phaser.Math.RND.between(0, this.physics.world.bounds.width);
     zone.y = Phaser.Math.RND.between(0, this.physics.world.bounds.height);
