+++ src/js/Scenes/PreloaderScene.js
@@ -85,7 +85,6 @@ export default class PreloaderScene extends Phaser.Scene {
 
     this.timedEvent = this.time.delayedCall(3000, this.ready, [], this);
 
-    // load assets needed in our game
     this.load.image('blueButton1', blueButton1);
     this.load.image('blueButton2', blueButton2);
     this.load.image('phaserLogo', phaserLogo);
@@ -94,7 +93,6 @@ export default class PreloaderScene extends Phaser.Scene {
     this.load.image('checkedBox', checkedBox);
     this.load.audio('bgMusic', [bgMusic]);
 
-    //---------------
   }
 
   ready() {
